Added separate MultiplicationMatrix class
